+++ src/main/java/ar/com/jumperinformatica/gui/controller/FacturaEgresoBController.java
@@ -2,6 +2,7 @@ package ar.com.jumperinformatica.gui.controller;
 
 import java.awt.Dialog;
 import java.awt.event.FocusListener;
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.text.ParseException;
 import java.util.Calendar;
@@ -58,7 +59,7 @@ public class FacturaEgresoBController extends FacturaController {
 		String numeroFactura = null;
 		Integer numeroRemito = null;
 		String obra  = null;
-		Float totalFactura  = null;
+		BigDecimal totalFactura  = null;
 		TipoIva tipoIva = null;
 	
 		NumberFormat locFormato = NumberFormat.getNumberInstance();
@@ -92,7 +93,7 @@ public class FacturaEgresoBController extends FacturaController {
 		
 		if (!this.vista.getTfTotal().getText().isEmpty()){
 			try{
-				totalFactura = locFormato.parse(this.vista.getTfTotal().getText()).floatValue();
+				totalFactura = (BigDecimal)locFormato.parse(this.vista.getTfTotal().getText());
 			} catch (ParseException e) {
 				e.printStackTrace();
 				throw new LogicaException("El total de la factura debe ser un número válido");
